Refactored out common DataUnit functionality from MultiContrastSegmentationEvaluationDataUnit and RegistrationReviewDataUnit
